made satellite and ship data accept parameters for input output and time

+++ set_grid_sizes.py
@@ -14,12 +14,14 @@ config_dir = conf_dat['config_dir']
 agt.read_cmd_params()
 
 for file_name in agt.model_parameters['files2newgrid']:
-    new_file_name =  re.sub('\.nc','_regridded.nc',file_name)
-    cdo_command = "cdo remapbil,{} {} {}".\
-            format(config_dir + agt.model_parameters['grid_template'],\
-                   oper_dir + file_name,\
-                   oper_dir + new_file_name)
-    os.system(cdo_command)
-    if(agt.model_parameters['del_temp_files'] and os.path.isfile(oper_dir + new_file_name)):
-            os.remove(oper_dir + file_name)
-
+    if(os.path.isfile(oper_dir + file_name)):
+        new_file_name =  re.sub('\.nc','_regridded.nc',file_name)
+        cdo_command = "cdo remapbil,{} {} {}".\
+                format(config_dir + agt.model_parameters['grid_template'],\
+                       oper_dir + file_name,\
+                       oper_dir + new_file_name)
+        os.system(cdo_command)
+        if(agt.model_parameters['del_temp_files'] and os.path.isfile(oper_dir + new_file_name)):
+                os.remove(oper_dir + file_name)
+    else:
+        print("{} not found.".format(oper_dir + file_name))
